[GridNG] Out of flow items with the parent of the grid as a containing block
This change fixes a test failure where the parent of the grid is the
containing block of an absolutely positioned child of the grid.
Previously, the containing block of out of flow children of the grid was
assumed to be the grid container. As a result, their static position of
was incorrect. Furthermore, this CL fixes the crash when adding a
position fixed item as child of the grid.

Bug: 1045599, 1190220
Change-Id: Iadb035783937bec117d594bb371ac989b08d9d1b