Remove Elements from toJSON
1. PerformanceElementTiming
2. LargestContentfulPaint
3. LayoutShiftAttribution

These include element or node in each toJSONForBinding implementation.

Those should be removed because they use the default toJSON operation and
those should not include interfaces that themselves do not have a toJSON
operation. So we should remove them from the toJSON builder and update
the existing toJSON WPT tests,

Bug: 1199954
Change-Id: I282439c17485264f1621102e3d002a693d48c1f9